refactor: add dependency of console module and move controllers into console module.
