--- eventmesh-dashboard-console/src/main/java/org/apache/eventmesh/dashboard/console/Main.java
@@ -1,7 +1,0 @@
-package org.apache.eventmesh.dashboard.console;
-
-public class Main {
-    public static void main(String[] args) {
-        System.out.println("Hello world!");
-    }
-}
